nrf_802154: rev 11986be74d191f6ea4de17d059cfb87ff7f86dd9
This commit updates revision of the nrf_802154 component to commit
11986be74d191f6ea4de17d059cfb87ff7f86dd9 of the sdk-nrf-802154
repository.

+++ nrf_802154/serialization/include/host/nrf_802154_callouts.h
@@ -125,7 +125,7 @@ extern void nrf_802154_tx_ack_started(const uint8_t * p_data);
 extern void nrf_802154_received_timestamp_raw(uint8_t * p_data,
                                               int8_t    power,
                                               uint8_t   lqi,
-                                              uint32_t  time);
+                                              uint64_t  time);
 
 /**
  * @brief Notifies that the reception of a frame failed.
