+++ nrf_802154/serialization/src/include/nrf_802154_spinel_dec_app.h
@@ -175,7 +175,7 @@ nrf_802154_ser_err_t nrf_802154_spinel_decode_prop_nrf_802154_capabilities_get_r
 nrf_802154_ser_err_t nrf_802154_spinel_decode_prop_nrf_802154_time_get_ret(
     const void * p_property_data,
     size_t       property_data_len,
-    uint32_t   * p_time);
+    uint64_t   * p_time);
 
 /**
  * @brief Decode SPINEL_PROP_VENDOR_NORDIC_NRF_802154_CCA_CFG_GET.
